% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghat-package.R
\docType{package}
\name{gghat}
\alias{gghat-package}
\alias{gghat}
\title{Adding hat graphs to the repertoire of 'ggplot2'}
\description{
The hat geom and compatible position functions for creating hat graphs in 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eliashyde/gghat}
  \item Report bugs at \url{https://github.com/eliashyde/gghat/issues}
}

}
\author{
\strong{Maintainer}: Elias Hyde \email{eliasworrallhyde@gmail.com}

}
\keyword{internal}
