% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-dodgedifferror.R
\name{position_dodgedifferror}
\alias{position_dodgedifferror}
\title{Dodged Difference for an Error Bar}
\usage{
position_dodgedifferror(width = NULL, backwards = FALSE)
}
\arguments{
\item{width}{Relative width of error bars.}

\item{backwards}{Whether or not to create backwards hats when the difference is negative.}
}
\value{
A \code{ggproto} object of class \code{PositionDodgeDiffError}.
}
\description{
Dodged Difference for an Error Bar
}
