% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-dodgedifftext.R
\name{position_dodgedifftext}
\alias{position_dodgedifftext}
\title{Dodged Difference for Text}
\usage{
position_dodgedifftext(width = NULL, nudge = 1, backwards = FALSE)
}
\arguments{
\item{width}{Relative width of bars and the like.}

\item{nudge}{Nudge direct labels up and down away from hat.}

\item{backwards}{Whether or not to create backwards hats when the difference is negative.}
}
\description{
Dodged Difference for Text
}
