% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-hat.R
\name{geom_hat}
\alias{geom_hat}
\title{The Hat Geom}
\usage{
geom_hat(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  backwards = FALSE,
  position = position_dodgediff(backwards = backwards),
  just = 0.5,
  width = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{As inherited.}

\item{data}{As inherited.}

\item{stat}{As inherited.}

\item{backwards}{Whether or not to create backwards hats when the difference is negative.}

\item{position}{As inherited.}

\item{just}{As inherited.}

\item{width}{As inherited.}

\item{na.rm}{As inherited.}

\item{show.legend}{As inherited.}

\item{inherit.aes}{As inherited.}
}
\description{
The hat geom functions much like a bar or col geom with stat = "identity" and
position = "dodge." It requires x, y, and group aesthetics, and it takes the
same parameters. However, the hat geom adopts "dodgediff" as its default position.
And as well as the parameters characteristic of bars, the hat geom may also be
passed a bool to direct its behavior in the case of negative difference, when
there is a decrease from a baseline for instance. This bool is the backwards
parameter described below.
}
