% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-dodgediff.R
\name{position_dodgediff}
\alias{position_dodgediff}
\title{The Dodged Difference Position}
\usage{
position_dodgediff(width = NULL, backwards = FALSE)
}
\arguments{
\item{width}{Relative width of bars and the like.}

\item{backwards}{Whether or not to create backwards hats when the difference is negative.}
}
\description{
The Dodged Difference Position
}
