% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-dodgedifftext.R
\name{position_dodgedifftext}
\alias{position_dodgedifftext}
\title{Dodged Difference for Text}
\usage{
position_dodgedifftext(width = 0.75, nudge = NULL, backwards = FALSE)
}
\arguments{
\item{width}{Relative width of text.}

\item{nudge}{Distance to move direct labels up and down away from hat.}

\item{backwards}{Whether or not to create backwards hats when the difference is negative.}
}
\value{
A \code{ggproto} object of class \code{PositionDodgeDiffText}.
}
\description{
Dodged Difference for Text
}
